Move motor code into individual includes. Update pinout. No longer functional...
